(creature_movement): grid-based movement with obstacles on walkable surfaces
